Wellborelar için : Derinlik, Kuyular için: Enlem, Boylam değerleri dolduruldu.

+++ ParsingProjectMVC/Models/WellboreModel.cs
@@ -2,14 +2,11 @@
 {
     public class WellboreModel
     {
-        
         public string WellboreAdi { get; set; }
-
         public int Id { get; set; }
-
-        public double? Derinlik { get; set; }
-        public SahaModel SahaAdi { get; set; } 
+        public int? Derinlik { get; set; } // Derinlik now int
+        public SahaModel SahaAdi { get; set; }
         public KuyuModel KuyuAdi { get; set; }
-        public KuyuGrubuModel KuyuGrubuAdi { get; set; }   
+        public KuyuGrubuModel KuyuGrubuAdi { get; set; }
     }
 }
